Rename back vitest.config.unit.mjs to vitest.config.mjs
