+++ src/Chartjs/assets/vitest.config.mjs
@@ -1,6 +1,6 @@
 import path from 'node:path';
 import { mergeConfig } from 'vitest/config';
-import configShared from '../../../vitest.config.unit.mjs';
+import configShared from '../../../vitest.config.mjs';
 
 export default mergeConfig(configShared, {
     test: {
