+++ src/React/assets/vitest.config.mjs
@@ -1,6 +1,6 @@
 import react from '@vitejs/plugin-react';
 import { mergeConfig } from 'vitest/config';
-import configShared from '../../../vitest.config.unit.mjs';
+import configShared from '../../../vitest.config.mjs';
 
 export default mergeConfig(configShared, {
     plugins: [react()],
