+++ src/TogglePassword/assets/vitest.config.mjs
@@ -1,4 +1,4 @@
 import { mergeConfig } from 'vitest/config';
-import configShared from '../../../vitest.config.unit.mjs';
+import configShared from '../../../vitest.config.mjs';
 
 export default mergeConfig(configShared, {});
